+++ src/main/java/com/jobdam/sns/entity/SnsMessage.java
@@ -16,10 +16,10 @@ public class SnsMessage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "message_id")
-    private Long messageId;
+    private Integer messageId;
 
     @Column(name = "receiver_user_id", nullable = false)
-    private Long receiverId;
+    private Integer receiverId;
 
     @Column(name = "sender_user_id", nullable = false)
     private Integer senderId;
